Fix structs for now.

+++ tests/float_test.fx
@@ -3,7 +3,7 @@
 def main() -> int
 {
     float a,b = 3.14,5;
-    print((a+b)/(float)3,5);
+    print((a+b)/3f,5);
     print();
     return 0;
 };
